requirements.txt
alembic==1.12.1
app==0.0.1
Babel==2.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
Mako==1.3.0
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
